requirements.txt
blinker==1.8.2
boto3==1.35.34
botocore==1.35.34
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.7.0
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mercadopago==2.2.3
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pymongo==4.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.35
stripe==11.1.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.3
Werkzeug==3.0.4